Added class Cannon to clean code and also added forceHit as an extra method for easier testing
