+++ src/Direction.java
@@ -3,5 +3,4 @@ public enum Direction {
   RIGHT,
   DOWN,
   LEFT,
-  UNKNOWN
 }
